only initializes user if in multirole layout

+++ src/VirtualClient/VirtualClient.Actions/SysbenchOLTP/SysbenchOLTPServerExecutor.cs
@@ -69,22 +69,25 @@ namespace VirtualClient.Actions
             string workingDirectory = this.GetPackagePath(this.PackageName);
             this.SystemManager.FileSystem.Directory.CreateDirectory(workingDirectory);
 
-            string dropUserCommand = $"mysql --execute=\"DROP USER IF EXISTS 'sbtest'@'{this.ClientIpAddress}'\"";
-            string configureNetworkCommand = $"sed -i \"s/.*bind-address.*/bind-address = {this.ServerIpAddress}/\" /etc/mysql/mysql.conf.d/mysqld.cnf";
-            string restartmySQLCommand = "systemctl restart mysql.service";
-            string createUserCommand = $"mysql --execute=\"CREATE USER 'sbtest'@'{this.ClientIpAddress}'\"";
-            string grantPrivilegesCommand = $"mysql --execute=\"GRANT ALL ON sbtest.* TO 'sbtest'@'{this.ClientIpAddress}'\"";
+            if (this.IsMultiRoleLayout()) 
+            {
+                string dropUserCommand = $"mysql --execute=\"DROP USER IF EXISTS 'sbtest'@'{this.ClientIpAddress}'\"";
+                string configureNetworkCommand = $"sed -i \"s/.*bind-address.*/bind-address = {this.ServerIpAddress}/\" /etc/mysql/mysql.conf.d/mysqld.cnf";
+                string restartmySQLCommand = "systemctl restart mysql.service";
+                string createUserCommand = $"mysql --execute=\"CREATE USER 'sbtest'@'{this.ClientIpAddress}'\"";
+                string grantPrivilegesCommand = $"mysql --execute=\"GRANT ALL ON sbtest.* TO 'sbtest'@'{this.ClientIpAddress}'\"";
 
-            await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(dropUserCommand, null, workingDirectory, cancellationToken)
-                    .ConfigureAwait(false);
-            await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(configureNetworkCommand, null, workingDirectory, cancellationToken)
-                    .ConfigureAwait(false);
-            await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(restartmySQLCommand, null, workingDirectory, cancellationToken)
-                    .ConfigureAwait(false);
-            await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(createUserCommand, null, workingDirectory, cancellationToken)
-                    .ConfigureAwait(false);
-            await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(grantPrivilegesCommand, null, workingDirectory, cancellationToken)
-                    .ConfigureAwait(false);
+                await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(dropUserCommand, null, workingDirectory, cancellationToken)
+                        .ConfigureAwait(false);
+                await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(configureNetworkCommand, null, workingDirectory, cancellationToken)
+                        .ConfigureAwait(false);
+                await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(restartmySQLCommand, null, workingDirectory, cancellationToken)
+                        .ConfigureAwait(false);
+                await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(createUserCommand, null, workingDirectory, cancellationToken)
+                        .ConfigureAwait(false);
+                await this.ExecuteCommandAsync<SysbenchOLTPServerExecutor>(grantPrivilegesCommand, null, workingDirectory, cancellationToken)
+                        .ConfigureAwait(false);
+            }
         }
     }
 }
